Set up basic persistent data
Right now, only buckles are carried over between runs (and game
startups). An option exists in the `Hub` script (`reset_save`) that
will reset save data on startup if enabled. Right now, it defaults to
`true`, but for the production build of the game, we'll want to make it
`false`.